Conflicts:
	src/core/TSQuery.java
	src/tools/CliQuery.java

+++ src/graph/Plot.java
@@ -76,7 +76,7 @@ public final class Plot {
    * Gnuplot always renders timestamps in UTC, so we simply apply a delta
    * to get local time.
    */
-  private final short utc_offset;
+  private final int utc_offset;
 
   /**
    * Constructor.
@@ -113,7 +113,7 @@ public final class Plot {
     if (tz == null) {
       tz = DEFAULT_TZ;
     }
-    this.utc_offset = (short) (tz.getOffset(System.currentTimeMillis()) / 1000);
+    this.utc_offset = tz.getOffset(System.currentTimeMillis()) / 1000;
   }
 
   /**
